requirements.txt
flask
bcrypt==4.1.1
python-dotenv
markdown
pytz